@@ -1,32 +1,21 @@
-import {ComponentType, FC} from "react";
-import {Redirect, Route} from "react-router";
+import {FC} from "react";
+import {Outlet} from "react-router";
+import {Navigate} from "react-router-dom";
 import WrapWithHeader from "../components/layout/WrapWithHeader";
 
-type authPathType = "/login" | "/register";
 
 interface PrivateRouteProps {
-  component: ComponentType,
-  path: authPathType,
-  exact?: boolean,
-  children?: React.ReactNode,
+
 }
 
-const AuthRoute: FC<PrivateRouteProps> = (
-  {
-    component,
-    children,
-    path,
-    exact,
-  }) => {
+const AuthRoute: FC<PrivateRouteProps> = () => {
   const isAuthenticated = localStorage.getItem('token') !== null;
 
   return isAuthenticated ? (
-    <Redirect to=""/>
+    <Navigate to=""/>
   ) : (
     <WrapWithHeader>
-      <Route path={path} exact={exact} component={component}>
-        {children}
-      </Route>
+      <Outlet/>
     </WrapWithHeader>
   );
 };
