@@ -7,7 +7,7 @@ const AxiosClient = axios.create({
 // before sending request attach auth token
 AxiosClient.interceptors.request.use((config) => {
   const token = localStorage.getItem('token');
-  config.headers.Authorization = token ? `Token ${token}` : '';
+  config!.headers!.Authorization = token ? `Token ${token}` : '';
   return config;
 });
 
